requirements.txt
streamlit
soundfile
torch
torchaudio
git+https://github.com/suno-ai/bark.git
ibm-watson-machine-learning
elevenlabs
elevenlabs==1.1.0
python-dotenv
